refactor: ingestion service
