+++ apps/api/src/index.ts
@@ -27,9 +27,9 @@ import { registerListPlanVersionsV1 } from "./routes/plans/listPlanVersionsV1"
 import { registerGetFeaturesV1 } from "./routes/project/getFeaturesV1"
 
 import { env } from "cloudflare:workers"
+import type { IngestionQueueMessage } from "@unprice/services/ingestion"
 import { timing } from "hono/timing"
 import { verifyRealtimeTicket } from "~/auth/ticket"
-import type { IngestionQueueMessage } from "~/ingestion/message"
 import { consumeIngestionBatch } from "~/ingestion/service"
 import { obs } from "~/middleware/obs"
 import { apiEvlog } from "~/observability"
