+++ internal/lakehouse/src/events-registry.test.ts
