+++ internal/services/package.json
@@ -20,6 +20,7 @@
     "./payment-provider": "./src/payment-provider/index.ts",
     "./cache": "./src/cache/index.ts",
     "./projects": "./src/projects/index.ts",
+    "./ingestion": "./src/ingestion/index.ts",
     "./apikey": "./src/apikey/index.ts",
     "./billing": "./src/billing/index.ts",
     "./entitlements": "./src/entitlements/index.ts",
@@ -47,6 +48,7 @@
     "@unprice/config": "workspace:^",
     "@unprice/db": "workspace:^",
     "@unprice/error": "workspace:^",
+    "@unprice/lakehouse": "workspace:^",
     "@unprice/logs": "workspace:^",
     "@unprice/metrics": "workspace:^",
     "@unprice/stripe": "workspace:^",
