+++ internal/services/src/entitlements/index.ts
@@ -4,3 +4,4 @@ export * from "./grants"
 export * from "./domain"
 export * from "./engine"
 export * from "./policy"
+export * from "./limit-policy"
